+++ tests/Dummies/Order.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Vanilo\Braintree\Tests\Dummies;
 
 use Illuminate\Database\Eloquent\Model;
+use Traversable;
 use Vanilo\Contracts\Billpayer;
 use Vanilo\Contracts\Payable;
 
@@ -40,5 +41,35 @@ class Order extends Model implements Payable
     public function getBillpayer(): ?Billpayer
     {
         return null;
+    }
+
+    public function getNumber(): string
+    {
+        return (string) $this->id;
+    }
+
+    public function getPayableRemoteId(): ?string
+    {
+        return null;
+    }
+
+    public function setPayableRemoteId(string $remoteId): void
+    {
+        // TODO: Implement setPayableRemoteId() method.
+    }
+
+    public static function findByPayableRemoteId(string $remoteId): ?Payable
+    {
+        return null;
+    }
+
+    public function hasItems(): bool
+    {
+        return false;
+    }
+
+    public function getItems(): Traversable
+    {
+        return collect();
     }
 }